Basic control flow for reverse mode
This is still very broken. I wasn't really planning to support this
until I put in some base compiler changes to make this easier, but
I think having it (at least at first order) is probably necessary
to get people to switch from Zygote, so let's put it in. It's not
particularly hard, just annoying. TBD how much of this we actually
want to do.