Import stage2 demand-driven forward mode
This imports a ripped-out version of the demand-driven AD code from
CedarSim and hooks it into the ADInterpreter. Starting with this code
has the advantage that it is working-ish, but the disadvantage that
it doesn't really interact with the rest of Diffractor yet. Still,
I think it's a reasonable point to start. I'm doing this as a
separate commit, so we can keep better track of the subsequent
refactoring.