Ensure that `:boundscheck` is properly handled
My previous fix was incorrect because `:boundscheck` returns a `Bool`.
Let's just force it to `true` for now, which implicitly disables
`@inbounds`, which is probably a good idea for us since we haven't gone
through the effort of ensuring that all indexing is still valid.
Add tests for this so that we catch this breakage faster next time.