stage1: Mark `partial` and `primal` helpers as `@inline`
It's unclear whether this is a compiler bug upstream, but I've observed
invokes like this in the wild:
```
%xxx = invoke DAECompiler.partial(%135::Diffractor.TangentBundle{1}, 1::Int64)::Any
```

which are (clearly) very poorly typed. `%135` was a fully concrete type,
so I'm not sure why the tangent type parameters ended up getting
stripped off.

In any case, the implementations here are so small that they are
probably best just inlined anyway.